% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{sentry_attach_lgr}
\alias{sentry_attach_lgr}
\title{Attach the Sentry \code{lgr} appender to a logger.}
\usage{
sentry_attach_lgr(logger = NULL, name = "sentry", threshold = NA_integer_)
}
\description{
Attach the Sentry \code{lgr} appender to a logger.
}
