% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_capture_log}
\alias{sentry_capture_log}
\title{Capture a structured Sentry log.}
\usage{
sentry_capture_log(level, message, logger, caller)
}
\description{
Capture a structured Sentry log.
}
