% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_capture_message}
\alias{sentry_capture_message}
\title{Capture a Sentry message at the requested level.}
\usage{
sentry_capture_message(message, level)
}
\description{
Capture a Sentry message at the requested level.
}
