% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_clear_current_span}
\alias{sentry_clear_current_span}
\title{Clear the active span from the current Sentry scope.}
\usage{
sentry_clear_current_span()
}
\description{
Clear the active span from the current Sentry scope.
}
