% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_init}
\alias{sentry_init}
\title{Initialize the global Sentry client.}
\usage{
sentry_init(
  dsn,
  environment,
  release,
  debug,
  shutdown_timeout_ms,
  traces_sample_rate
)
}
\description{
Initialize the global Sentry client.
}
