% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{sentry_initialize}
\alias{sentry_initialize}
\title{Initialize the global Sentry client.}
\usage{
sentry_initialize(
  dsn,
  environment = NULL,
  release = NULL,
  debug = FALSE,
  shutdown_timeout_ms = 2000L,
  traces_sample_rate = 1
)
}
\description{
Initialize the global Sentry client.
}
