% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{sentry_lgr_appender}
\alias{sentry_lgr_appender}
\title{Create an \code{lgr} appender that forwards logs to Sentry.}
\usage{
sentry_lgr_appender(threshold = NA_integer_)
}
\description{
Create an \code{lgr} appender that forwards logs to Sentry.
}
