% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_logs_enabled}
\alias{sentry_logs_enabled}
\title{Report whether the global Sentry client is enabled for structured logs.}
\usage{
sentry_logs_enabled()
}
\description{
Report whether the global Sentry client is enabled for structured logs.
}
