% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{sentry_plumber_instrument}
\alias{sentry_plumber_instrument}
\title{Instrument a plumber router with Sentry request tracing.}
\usage{
sentry_plumber_instrument(
  pr,
  transaction_namer = NULL,
  flush_timeout_ms = 2000L,
  error_handler = NULL
)
}
\description{
Instrument a plumber router with Sentry request tracing.
}
