% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_set_current_span}
\alias{sentry_set_current_span}
\title{Set the active span on the current Sentry scope.}
\usage{
sentry_set_current_span(span)
}
\description{
Set the active span on the current Sentry scope.
}
