% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_span_start_child}
\alias{sentry_span_start_child}
\title{Start a child span from an existing transaction or span.}
\usage{
sentry_span_start_child(parent, op, description)
}
\description{
Start a child span from an existing transaction or span.
}
