% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sentry_transaction_start}
\alias{sentry_transaction_start}
\title{Start a transaction, optionally continuing an inbound \code{sentry-trace} header.}
\usage{
sentry_transaction_start(name, op, sentry_trace)
}
\description{
Start a transaction, optionally continuing an inbound \code{sentry-trace} header.
}
