% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{with_sentry_transaction}
\alias{with_sentry_transaction}
\title{Run code inside a Sentry transaction.}
\usage{
with_sentry_transaction(name, op, code, sentry_trace = NULL)
}
\description{
Run code inside a Sentry transaction.
}
